+++ pkg/virt-launcher/virtwrap/device/hostdevice/sriov/pcipool_test.go
@@ -27,7 +27,7 @@ import (
 	. "github.com/onsi/gomega"
 
 	v1 "kubevirt.io/client-go/api/v1"
-	"kubevirt.io/kubevirt/pkg/virt-launcher/virtwrap/device/sriov"
+	"kubevirt.io/kubevirt/pkg/virt-launcher/virtwrap/device/hostdevice/sriov"
 )
 
 type networkData struct {
