+++ pkg/virt-launcher/virtwrap/device/hostdevice/sriov/sriov_suite_test.go
@@ -26,7 +26,7 @@ import (
 	"kubevirt.io/client-go/testutils"
 )
 
-func TestNetwork(t *testing.T) {
+func TestSRIOV(t *testing.T) {
 	testutils.KubeVirtTestSuiteSetup(t)
 }
